requirements.txt
pandas==2.2.3
python-dotenv==1.1.0
influxdb-client==1.48.0
streamlit==1.32.0
streamlit-aggrid==1.1.3
altair==5.2.0
plotly==5.22.0